[FLINK-2781] [core] Cleanup NetUtils.
  - The NetUtils class (in flink-core) contains all methods usable without runtime dependency
  - The runtime NetUtils class (to find connectiong addresses) is now called ConnectionUtil.
